% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{clean_likert_numeric_vector}
\alias{clean_likert_numeric_vector}
\title{Clean Likert Numeric Vector}
\usage{
clean_likert_numeric_vector(raw_vector)
}
\arguments{
\item{raw_vector}{A numeric vector containing raw Likert scale responses.}
}
\value{
A numeric vector where each element is scaled to a 0-1 range.
}
\description{
Transforms raw Likert scale responses into a normalized numeric vector.
The transformation scales the data to a 0-1 range, based on the number of levels in the Likert scale.
This is particularly useful when Likert scale responses need to be treated as continuous variables
for statistical analysis.
}
\examples{
# Example data
raw_data <- c(1, 2, 3, 4, 5, 1, 2, 3, 4, 5)

# Clean the data using clean_likert_numeric_vector
clean_data <- clean_likert_numeric_vector(raw_data)
print(clean_data)

# Applying it to a data frame column example
data_clean$comp_sante_focus_attention_present <- clean_likert_numeric_vector(data_raw$autogestion_1)

}
