% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging.R
\name{extract_elements_with_prefix}
\alias{extract_elements_with_prefix}
\title{Extract elements with names starting with a specific prefix}
\usage{
extract_elements_with_prefix(vec, prefix)
}
\arguments{
\item{vec}{A named vector.}

\item{prefix}{A string representing the prefix to match.}
}
\value{
A named vector containing the elements whose names start with the given prefix.
}
\description{
This function extracts elements from a vector whose names start with a given prefix.
}
\examples{
vec <- c(foo1 = 1, foo2 = 2, bar1 = 3, bar2 = 4)
extract_elements_with_prefix(vec, "foo")
}
