% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{finverser}
\alias{finverser}
\title{Invert the Order of Unique Values in a Vector}
\usage{
finverser(vec_col)
}
\arguments{
\item{vec_col}{A numeric vector containing the values to be inverted.}
}
\value{
A numeric vector of the same length as `vec_col`, with the order of unique values inverted.
        The spacing between different values is preserved in the inverted order.
}
\description{
This function inverts the order of unique numerical values in a vector. For each unique value in the input vector,
it assigns a new value that maintains the original spacing but in reverse order.
It's useful for transforming data in a way that preserves the relative magnitude of values while reversing their order.
}
\examples{
finverser(c(1, 2, 3, 2, 1)) # Returns c(3, 2, 1, 2, 3)
finverser(c(10, 20, 30))    # Returns c(30, 20, 10)

}
