% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging.R
\name{generate_survey_ids}
\alias{generate_survey_ids}
\title{Generate Survey IDs}
\usage{
generate_survey_ids(n_respondents, source_id)
}
\arguments{
\item{n_respondents}{The number of respondents, i.e., the length of the sequence.}

\item{source_id}{A character string representing the source ID (survey).}
}
\value{
A character vector containing the generated survey IDs.
}
\description{
This function generates survey IDs by combining a source ID with a sequence of numbers.
}
\details{
This function generates survey IDs by combining a source ID with a sequence of numbers.
For example, if the source ID is "ces65" and n_respondents is 3, the function will return
the vector "ces65.1", "ces65.2", "ces65.3".
}
\examples{
generate_survey_ids(n_respondents = 5, source_id = "ces65")

}
