% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging.R
\name{load_variable}
\alias{load_variable}
\title{Load variable from file}
\usage{
load_variable(file, variable_name)
}
\arguments{
\item{file}{A string containing the path to a file.}

\item{variable_name}{A string containing the name of the variable.}
}
\value{
A vector with the same number of rows as the file containing the variable values
}
\description{
This function takes a file and a variable name and returns the vector
containing the variable values.
}
\examples{
# Create a temporary CSV file with example data
temp_file <- tempfile(fileext = ".csv")
writeLines(c("var1,var2,var3", "1,2,3", "4,5,6", "7,8,9"), temp_file)

# Load the variable 'var2' from the temporary CSV file
load_variable(file = temp_file, variable_name = "var2")
}
