% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging.R
\name{match_and_update}
\alias{match_and_update}
\title{Update a Main Vector with Updates by Matching Names}
\usage{
match_and_update(main, updates)
}
\arguments{
\item{main}{A vector to update.}

\item{updates}{A vector containing the updates to main.}
}
\value{
The updated \code{main} vector with values from the \code{updates} vector where names match.
}
\description{
This function updates a main vector with values from an updates vector by matching names.
It replaces values in the main vector with corresponding values from the updates vector
where the names are the same.
}
\examples{
main <- c(a = 1, b = 2, c = 3, d = 4)
updates <- c(b = 22, c = 33)
updated_vector <- match_and_update(main, updates)

}
