% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{minmaxNormalization}
\alias{minmaxNormalization}
\title{Min-Max Normalization}
\usage{
minmaxNormalization(x)
}
\arguments{
\item{x}{A numeric vector to be normalized.}
}
\value{
A numeric vector with values scaled between 0 and 1.
}
\description{
This function performs min-max normalization on a numeric vector. 
It scales the values to a range between 0 and 1.
}
\examples{
vec <- c(1, 2, 3, 4, 5)
minmaxNormalization(vec)
# [1] 0.00 0.25 0.50 0.75 1.00
}
