% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{parse_money_range}
\alias{parse_money_range}
\title{Parse a Money Range}
\usage{
parse_money_range(value, sep = NULL, limit = NULL)
}
\arguments{
\item{value}{A character string containing the range of monetary values.}

\item{sep}{A character string containing the separator between the lower and upper bounds of the range. This parameter is required if `limit` is `NULL`.}

\item{limit}{A character string specifying how to handle values that do not have an explicit upper or lower bound. If `limit` is `"floor"`, the function assumes a lower bound of 0. If `limit` is `"ceiling"`, the function assumes an upper bound that is 30,000 units higher than the specified value. If `limit` is `NULL`, the function assumes that the value contains both lower and upper bounds, separated by `sep`.}
}
\value{
A numeric vector of length two, where the first element is the lower bound and the second element is the upper bound of the range.
}
\description{
This function parses a string containing a range of monetary values and returns a numeric vector of length two, where the first element is the lower bound and the second element is the upper bound of the range.
}
\examples{
parse_money_range("$1,000 to $1,999", sep = "to")
parse_money_range("under $1,000", limit = "floor")
parse_money_range("$15,000 or over", limit = "ceiling")

}
