% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{qualtrics_na_counter}
\alias{qualtrics_na_counter}
\title{Check and Report Missing Values in DataFrame Columns}
\usage{
qualtrics_na_counter(data)
}
\arguments{
\item{data}{A DataFrame containing the data to be analyzed for missing values.}
}
\value{
A DataFrame with each variable group, the count of NAs, and the percentage of NAs.
Additionally, a bar plot is displayed showing the percentage of missing data for each variable group
with non-zero missing data.
}
\description{
This function processes a DataFrame, typically sourced from a Qualtrics .sav file, to identify and quantify
missing values across all columns. It specifically treats empty strings in text columns as NA, recalculates
missing values, and then groups variables by their names minus numeric and text suffixes to calculate the
count and percentage of missing data per group. It also generates a bar plot displaying the percentage of
missing data for variable groups where the percentage is greater than zero.
}
\examples{
data <- data.frame(
  ID_1_TEXT = c("apple", "", "banana"),
  Value_1 = c(1, 2, NA),
  ID_2_TEXT = c("", "orange", "melon"),
  Value_2 = c(3, 4, 5)
)
results <- check_na(data)
print(results)

}
