% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging.R
\name{read_any_csv}
\alias{read_any_csv}
\title{Robust CSV Reading Function}
\usage{
read_any_csv(file_path, ...)
}
\arguments{
\item{file_path}{A string specifying the file path and name.}

\item{...}{Additional arguments to be passed to the reading functions.}
}
\value{
A data frame containing the data read from the CSV file.
}
\description{
This function works through a series of attempts to read a CSV file, employing
different strategies to navigate around typical issues that might cause
reading functions to fail. It uses both base R and `readr` functions, trying
different combinations of parameters to facilitate reading in different scenarios.
}
\details{
`read_any_csv()` aims to robustly read a CSV file, navigating through common issues
such as different delimiters, problematic headers, and others. It employs various reading
methods in a series of attempts to successfully read the file.


The function proceeds through a series of attempts to read the CSV file:
  1. Standard reading with `read.csv`.
  2. Reading without headers using `read.csv`.
  3. Using utils::read.csv2
  4. Utilizing `readr::read_csv`.
  5. Reading with `read.table` and `fill = TRUE`.
If all attempts fail, the function stops and returns an error message.
}
\section{Handling Issues}{

- If headers are problematic or have a different number of column names,
  it tries reading without headers.
- For differences in managing CSV files between `read.csv` and `read_csv`,
  it tries both.
- It attempts to fill empty columns, which can be problematic in some instances.
}

\examples{
\dontrun{
data <- read_any_csv("path/to/your/file.csv")
}

}
