% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging.R
\name{read_survey}
\alias{read_survey}
\title{Read Survey Data}
\usage{
read_survey(file)
}
\arguments{
\item{file}{A character string specifying the name of the file to be read. The file should have either a .csv, .xlsx, or .Sav file extension.}
}
\value{
A data frame containing the contents of the file. The data type (e.g., character, numeric) of each column in the returned data frame will depend on the file content.
}
\description{
This function reads in survey data from a specified file, handling various file formats (csv, xlsx, and Sav).
}
\note{
Ensure that the appropriate libraries (readxl, haven) are installed before using this function to read .xlsx or .Sav files respectively.
}
\examples{
\dontrun{
  data_csv <- read_survey("path/to/your/file.csv")
  data_xlsx <- read_survey("path/to/your/file.xlsx")
  data_sav <- read_survey("path/to/your/file.Sav")
}

}
\seealso{
\code{\link[utils]{read.csv2}}, \code{\link[readxl]{read_xlsx}}, \code{\link[haven]{read_sav}} for the functions used to read data based on file type.
}
