% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{rename_snakecase}
\alias{rename_snakecase}
\title{Rename Data Frame Columns to Snake Case}
\usage{
rename_snakecase(df)
}
\arguments{
\item{df}{A data frame whose columns are to be renamed.}
}
\value{
A data frame with the same structure as the input, but with snake_case
        column names.
}
\description{
This function renames the columns of a data frame to snake_case.
It uses the \code{\link{dplyr::rename_with}} function and a regular expression to
insert underscores between words.
}
\examples{
df <- data.frame(FirstName = c("John", "Jane"), LastName = c("Doe", "Smith"))
rename_snakecase(df)
#   first_name last_name
# 1       John       Doe
# 2       Jane     Smith

}
