% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{topdown_fa}
\alias{topdown_fa}
\title{Top-Down Factor Analysis}
\usage{
topdown_fa(df, nfactors = 1)
}
\arguments{
\item{df}{A data frame where each column represents a variable to be included in the factor analysis.}

\item{nfactors}{The number of factors to extract. Default is 1.}
}
\value{
A ggplot object showing the factor loadings plot with annotations for Cronbach's alpha and the first eigenvalue. The function also invisibly returns a list containing Cronbach's alpha, the first eigenvalue, and the factor loadings.
}
\description{
Performs a top-down factor analysis on a given data frame, computes Cronbach's alpha for reliability, conducts a factor analysis to extract the loadings of the first factor, and plots the factor loadings along with annotations for Cronbach's alpha and the first eigenvalue.
}
\examples{
# Example usage:
# df <- data.frame(variable1 = rnorm(100), variable2 = rnorm(100))
# topdown_fa(df)
}
