% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_data.R
\docType{data}
\name{us_counties_data}
\alias{us_counties_data}
\title{US Counties Data}
\format{
A data frame with X rows and Y variables:
\describe{
  \item{county_fips}{FIPS code of the county}
  \item{population_density}{Population density in square miles}
  \item{vote_projection_democrat}{ARIMA projection of democrat vote share in 2024}
}
}
\usage{
data(us_counties_data)
}
\description{
A dataset containing information about US counties, such as area and other attributes.
}
\keyword{datasets}
