% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging.R
\name{load_variable}
\alias{load_variable}
\title{Load variable from file}
\usage{
load_variable(file, variable_name)
}
\arguments{
\item{file}{A string containing the path to a file.}

\item{raw_variable_name}{A string containing the name of the variable.}
}
\value{
A vector with the same number of rows as the file containing the variable values
}
\description{
This function takes a file and a variable name and returns the vector
containing the variable values.
}
\examples{
load_variable(file = "data.csv", variable_name = "var332") ## should return the vector containing the values of var332.
}
